Path.untemp now removes read-only flag:
See:
https://afkmods.iguanadons.net/index.php?/topic/4966-wrye-bash-all-games/&do=findComment&comment=172227

WindowsError: [Error 5] Access is denied: u'D:\\Documents\\My Games\\FalloutNV\\Fallout.ini'

See following posts for strange customs of Fallout.ini. Was read only
so on setting Archive invalidation it blew. Fix in a central place,
still internal paths API is a mess - see #241.

+++ Mopy/bash/bolt.py
@@ -31,6 +31,7 @@ import collections
 import copy
 import csv
 import datetime
+import errno
 import gettext
 import locale
 import os
@@ -1010,8 +1011,15 @@ class Path(object):
                 else:
                     # this will fail with Access Denied (!) if self._s is
                     # (unexpectedly) a directory
-                    os.remove(self._s)
+                    try:
+                        os.remove(self._s)
+                    except OSError as e:
+                        if e.errno != errno.EACCES:
+                            raise
+                        self.clearRO()
+                        os.remove(self._s)
             shutil.move(self.temp._s, self._s)
+
     def editable(self):
         """Safely check whether a file is editable."""
         delete = not os.path.exists(self._s)
